% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bvalue_stress}
\alias{bvalue_stress}
\title{Differential stress from b-values based on Scholz's regression}
\usage{
bvalue_stress(b = 1)
}
\arguments{
\item{b}{numeric; earthquakes b-value}
}
\description{
Differential stress from b-values based on Scholz's regression
}
\details{
Differential stress (S1 - S3) may be inversely proportional to b-value Scholz (2015).
}
\examples{
bvalue_stress() # 1.0 characteristic of So CA: 192 MPa
bvalue_stress(c(0.85,0.90,1.0))
}
