% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hardebeck.R
\name{dTheta}
\alias{dTheta}
\title{Rotations in S1}
\usage{
dTheta(theta.deg = 30, dTau = 1, Tau = 1)
}
\arguments{
\item{theta.deg}{numeric;}

\item{dTau}{numeric;}

\item{Tau}{numeric;}
}
\description{
This calculates the
change in orientation of the maximum principal stress relative to
the fault prior to the earthquake (theta)
for a given stress drop (dTau) relative to pre-mainshock levels (Tau).
}
\examples{
thet.=seq(-90,90,by=2)
rats <- seq(-1,1,by=0.2)
All.dthet <- plyr::ldply(rats, function(r) dTheta(thet., dTau=r))
matplot(x=thet., t(All.dthet), type='l', col=c(5:1,NA,1:5), lty=c(1:5,1,5:1))
}
\references{
Hardebeck and Hauksson (2001): \url{http://dx.doi.org/10.1029/2001JB000292}
}

