% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticity.R
\name{elastic_compliance_tensor}
\alias{elastic_compliance_tensor}
\title{Calculate the isotropic elasticity tensor using Hooke's law of elasticity}
\usage{
elastic_compliance_tensor(nu = 0.25, Youngs = 7.5e+10)
}
\arguments{
\item{nu}{numeric; the Poisson's ratio}

\item{Youngs}{numeric; Young's modulus}
}
\value{
A symmetric matrix of elastic constants
}
\description{
Calculate the isotropic elasticity tensor using Hooke's law of elasticity
}
\details{
Hooke's law is relating stress \eqn{\sigma} to strain \eqn{\epsilon} is
\deqn{\sigma = \strong{C} \epsilon}
or 
\deqn{\sigma = 2 \mu \epsilon + \lambda \strong{I} \diag(\epsilon)}
where \eqn{\lambda,\mu} are Lame's first and second constants of elasticity.

This function calculates \eqn{C} for 3D elasticity -- a 6x6 tensor with three independent quantities.
}
\examples{
elastic_compliance_tensor()

}
\seealso{
\code{\link{youngs-mod}}
}
