% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msatsi.R
\name{msatsi-methods}
\alias{msatsi-methods}
\alias{msatsi}
\alias{print.msatsi_slboot_trpl}
\alias{print.msatsi_mat}
\alias{print.msatsi_slboot_tensor}
\alias{print.msatsi_summary}
\alias{print.msatsi_summary_ext}
\alias{print.msatsi_input}
\alias{summary.msatsi_input}
\alias{print.msatsi_input.summary}
\alias{plot.msatsi_input.summary}
\title{MSATSI methods}
\usage{
\method{print}{msatsi_slboot_trpl}(x, ...)

\method{print}{msatsi_mat}(x, ...)

\method{print}{msatsi_slboot_tensor}(x, ...)

\method{print}{msatsi_summary}(x, ...)

\method{print}{msatsi_summary_ext}(x, ...)

\method{print}{msatsi_input}(x, ...)

\method{summary}{msatsi_input}(object, ...)

\method{print}{msatsi_input.summary}(x, ...)

\method{plot}{msatsi_input.summary}(x, ...)
}
\arguments{
\item{x, object}{an object}

\item{...}{additional parameters, usually passed to
\code{\link{summary}},
\code{\link{print}}, or
\code{\link{plot}}}
}
\description{
The functions here can be used
to manipulate classes based on
MSATSI input and output files.
}
\seealso{
\code{\link{read_msatsi}}
}
