% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsm.R
\name{read.wsm}
\alias{read.wsm}
\title{Read data generated by the World Stress Map API}
\usage{
read.wsm(request.id, data.dir = file.path(request.id), ...)
}
\arguments{
\item{request.id}{character; the id assigned to the dataset}

\item{data.dir}{character; the location of the data from \code{request.id}}

\item{...}{additional parameters to \code{\link{read.table}}}
}
\description{
Read data generated by the World Stress Map API
}
\details{
Stress regime defined in the following manner:
TF: thrust faulting SH>Sh>SV
TS: thrust with strike-slip component
SS: strike-slip SH>SV>Sh
NS: normal with strike-slip component
NF: normal faulting SV>SH>Sh
U: unknown

Type of plate boundary according to the global plate model PB2002 from Bird (2003):
CTF=Continental Transform Fault
CRB=Continental Rift Boundary
CCB=Continental Collision Boundary
OTF=Oceanic Transform Fault
OSR=Oceanic Spreading Ridge
OCB=Oceanic Collision Boundary
SUB=Subduction zone
}
\references{
Heidbach, O., Tingay, M., Barth, A., Reinecker, J., Kurfess, D., and Mueller, B. (2008),
\emph{The World Stress Map database release 2008},
\url{http://dx.doi.org/10.1594/GFZ.WSM.Rel2008}

Heidbach, O., and J. Hohne (2008),
\emph{CASMI - a tool for the visualization of the World Stress Map data base},
Computers and Geosciences, 34 (7), 783-791,
\url{http://dx.doi.org/10.1016/j.cageo.2007.06.004}
}
