% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/msatsi.R
\name{relative_stress}
\alias{relative_stress}
\alias{relative_stress.slboot_tensor}
\title{Return the relative stress magnitudes and directions}
\usage{
relative_stress(x, ...)

\method{relative_stress}{slboot_tensor}(x, ...)
}
\description{
Return the relative stress magnitudes and directions
}
\seealso{
\code{\link{read_msatsi}}
}

