% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticity.R
\name{seismic_elasticity}
\alias{seismic_elasticity}
\title{Calculate elastic properties from a velocity model}
\usage{
seismic_elasticity(vp, vs, rho)
}
\arguments{
\item{vp}{numeric; the P-wave speed [m/s]}

\item{vs}{numeric; the S-wave speed [m/s]}

\item{rho}{numeric; the density [kg/m^3]}

\item{...}{additional parameters (e.g., passing \code{nu})}
}
\value{
bulk modulus, shear modulus, Lame's first constant;
in Pa if SI given for inputs
}
\description{
Calculate elastic properties from a velocity model
}
\examples{
# inputs in m/s and kg/m^3 give results in Pascal
Pa <- seismic_elasticity(7200, 3500, 2650)

# use km/s and g/cm^3 instead to get giga-Pascal
GPa <- seismic_elasticity(7.2, 3.5, 2.65)
all.equal(Pa/1e9, GPa)
}
