% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lund-sh.R
\name{SH}
\alias{SH}
\title{Calculation of the direction of maximum horizontal stress using a 3D relative stress tensor}
\usage{
SH(R, n1, n2 = NULL, n3 = NULL)
}
\arguments{
\item{R}{numeric; the principal stress ratio; defined as (S1 - S2)/(S1 - S3)}

\item{n1}{numeric; the unit vector of the maximum principal stress, S1 (North,East,Up); if
either \code{n2} or \code{n3} are \code{NULL} then this may be a three-column matrix}

\item{n2}{numeric; unit vector of the intermediate principal stress, S2}

\item{n3}{numeric; unit vector of the minimum principal stress, S3}
}
\value{
numeric; the azimuth of SHmax (c.w. from North) in radians
with full or partial knowledge of the tectonic stress tensor,
Geophys. J. Int., 170, 1328-1335, doi: 10.1111/j.1365-246X.2007.03468.x
}
\description{
Calculation of the direction of maximum horizontal stress using a 3D relative stress tensor
}
\examples{
# Dummy examples
Sf <- matrix(1:9, 3)
Sf[lower.tri(Sf)] = t(Sf)[lower.tri(Sf)]

S <- eigen(Sf)
principal_stresses <- S[['values']]
orientations <- S[['vectors']]

SH(0.5, orientations[,1], orientations[,2], orientations[,3])
SH(0.5, orientations)

Rseq <- seq(0, 1, length.out=21)
Azims <- sapply(Rseq, SH, n1=orientations)
plot(Rseq, Azims*180/pi, type='b')

}
\references{
Equations 10 and 11 from Lund and Townend, (2007): Calculating horizontal stress orientations
}
\author{
Copyright (C) 1998-2007 Bjorn Lund
}
