% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{shape_factor}
\alias{shape_factor}
\title{Stress ellipsoid, or 'shape factor'}
\usage{
shape_factor(S1, S2, S3, check.order = TRUE, stress.ratio = FALSE)
}
\arguments{
\item{S1}{numeric; the maximum principal stress}

\item{S2}{numeric; the intermediate principal stress}

\item{S3}{numeric; the minimum principal stress}

\item{check.order}{logical; should the order of the stress be checked?
If the check fails the indices which caused it to fail will be returned.
Set this to \code{FALSE} if you want to reduce computation time and
are confident in the order of \code{S1}, \code{S2}, and \code{S3}.}

\item{stress.ratio}{logical; should the \code{\link{stress_ratio}}
be returned instead?}
}
\description{
Stress ellipsoid, or 'shape factor'
}
\details{
If \code{S1} > \code{S2} > \code{S3} then:
\deqn{
 \phi = \frac{S2 - S3}{S1 - S3}
}
which is the parameter returned by SATSI, for example.

Note that the \code{\link{stress_ratio}} is:
\deqn{
 R = 1 - \phi = \frac{S1 - S2}{S1 - S3}
}
}

