% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stress-package.R
\docType{package}
\name{stress}
\alias{stress}
\alias{stress-package}
\title{The \code{stress} package: working with stress inversion results}
\description{
The \code{stress} package: working with stress inversion results
}
\seealso{
\code{\link{dTheta}}
}

