% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{stress_ellipsoid}
\alias{stress_ellipsoid}
\title{Stress ellipsoid, or shape factor}
\usage{
stress_ellipsoid(S1, S2, S3)
}
\arguments{
\item{S1}{numeric; the maximum principal stress}

\item{S2}{numeric; the intermediate principal stress}

\item{S3}{numeric; the minimum principal stress}
}
\description{
Stress ellipsoid, or shape factor
}

