% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{stress_ratio}
\alias{stress_ratio}
\title{Principal stress ratio}
\usage{
stress_ratio(S1, S2, S3, ...)
}
\arguments{
\item{S1}{numeric; the maximum principal stress}

\item{S2}{numeric; the intermediate principal stress}

\item{S3}{numeric; the minimum principal stress}

\item{...}{additional parameters to \code{\link{shape_factor}}}
}
\description{
Principal stress ratio
}
\seealso{
\code{\link{shape_factor}}
}

