% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spatial.R
\name{to_raster}
\alias{to_raster}
\alias{to_raster.msatsi_summary_ext}
\title{Convert objects to raster objects}
\usage{
to_raster(x, ...)

\method{to_raster}{msatsi_summary_ext}(x, transpose = FALSE,
  make.360 = TRUE, ...)
}
\arguments{
\item{x}{object}

\item{...}{additional parameters to
\code{\link[raster]{rasterFromXYZ}}}

\item{transpose}{logical; should the x and y coordinates be swapped?}

\item{make.360}{logical; should trends be in the range [0,360] rather than [-180,180]?}
}
\description{
Convert objects to raster objects
}

