% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{TP}
\alias{TP}
\title{Stereonet representation of trend and plunge}
\usage{
TP(ang.trend, ang.plunge, r = NULL, add = FALSE, pch = 16, ...)
}
\arguments{
\item{ang.trend}{numeric; trend or azimuth}

\item{ang.plunge}{numeric; plunch or dip}

\item{r}{numeric; the radius of the stereonet}

\item{add}{logical; should the points be added to the current device?}

\item{pch}{numeric; the symbol type of the points}

\item{...}{additional arguments to \code{\link{points}}}
}
\description{
Uses the equal-area (lower) hemispherical, or \emph{Schmidt}, projection.
 From [1]:  The main advantage of this projection
 method is that it does not suffer from the areal distortion of the equal angle
 projection. This means, for example, that an area such as that enclosed by a circle
 of constant radius on the projection, represents the same amount of area on the
 reference sphere regardless of its position. The circle will not however represent a
 geometrically similar circle on the sphere unless it is positioned at the centre of the
 projection. In other words, areal relationships are preserved in this method while
 geometrical relationships are distorted; the opposite is true of the equal angle
 projection.
}
\details{
Uses \code{\link{complex}}, which recycles input vectors as needed
}
\examples{
TP(runif(20, min = 0, max = 90),c(45,55))
TP(runif(20, min = 180, max = 270),c(45,55), pch=1, add=TRUE)
}
\references{
[1] Diederichs, Mark S (1990),
Dips: An Interactive and Graphical Approach to the Analysis of Orientation Based Data,
\url{https://www.rocscience.com/documents/pdfs/uploads/7672.pdf}
}

