% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticity.R
\name{youngs-mod}
\alias{youngs-mod}
\alias{shear_to_youngs}
\alias{bulk_to_youngs}
\alias{shear_to_bulk}
\alias{bulk_to_shear}
\title{Young's modulus and elastic constants}
\usage{
shear_to_youngs(mu, nu = 0.25)

bulk_to_youngs(kap, nu = 0.25)

shear_to_bulk(mu, nu = 0.25)

bulk_to_shear(kap, nu = 0.25)
}
\arguments{
\item{mu}{numeric; the elastic shear modulus}

\item{nu}{numeric; the Poisson's ratio}

\item{kap}{numeric; the elastic bulk modulus}
}
\examples{

# Calculate Young's modulus from

# shear modulus
sm <- 30e9
ym.s <- shear_to_youngs(sm)

# bulk modulus
bm <- shear_to_bulk(sm)
ym.b <- bulk_to_youngs(bm)

all.equal(ym.s, ym.b)

smi <- bulk_to_shear(bm)
all.equal(sm, smi)

}
\references{
\url{https://en.wikipedia.org/wiki/Elastic_modulus} for various inter-relationships
}
\seealso{
\code{\link{elastic_compliance_tensor}}
}
