% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{dip_direction}
\alias{dip_direction}
\title{Calculate dip direction from strike}
\usage{
dip_direction(strike = 0)
}
\arguments{
\item{strike.deg}{numeric; the strike of the nodal plane, in degrees}
}
\description{
Calculate dip direction from strike
}

