% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/msatsi.R
\name{read_msatsi}
\alias{plot.msatsi_input.summary}
\alias{print.msatsi_input}
\alias{print.msatsi_input.summary}
\alias{print.msatsi_slboot_tensor}
\alias{print.msatsi_slboot_trpl}
\alias{print.msatsi_summary}
\alias{print.msatsi_summary_ext}
\alias{read.msatsi_mat}
\alias{read.msatsi_sat}
\alias{read.msatsi_slboot_tensor}
\alias{read.msatsi_slboot_trpl}
\alias{read.msatsi_summary}
\alias{read.msatsi_summary_ext}
\alias{read_msatsi}
\alias{summary.msatsi_input}
\title{MSATSI output readers}
\usage{
read.msatsi_slboot_trpl(fi, ...)

\method{print}{msatsi_slboot_trpl}(x, ...)

read.msatsi_mat(fi, ...)

\method{print}{msatsi_slboot_trpl}(x, ...)

read.msatsi_slboot_tensor(fi, ...)

\method{print}{msatsi_slboot_tensor}(x, ...)

read.msatsi_summary(fi, ...)

\method{print}{msatsi_summary}(x, ...)

read.msatsi_summary_ext(fi, ...)

\method{print}{msatsi_summary_ext}(x, ...)

read.msatsi_sat(fi, ...)

\method{print}{msatsi_input}(x, ...)

\method{summary}{msatsi_input}(object, ...)

\method{print}{msatsi_input.summary}(x, ...)

\method{plot}{msatsi_input.summary}(x, ...)
}
\arguments{
\item{fi}{character; the filename}

\item{...}{additional parameters, usually passed to
\code{\link{read.table}},
\code{\link{print}},
\code{\link[R.matlab]{readMat}}, or
\code{\link{plot}}}

\item{x,object}{an object}
}
\description{
MSATSI has a standardized set of output files.  These functions
can be used to read-in certain files and create associated
objects.
}

