% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/msatsi.R
\name{read_msatsi}
\alias{plot.msatsi_input.summary}
\alias{read.msatsi_mat}
\alias{read.msatsi_sat}
\alias{read.msatsi_slboot_tensor}
\alias{read.msatsi_slboot_trpl}
\alias{read.msatsi_summary}
\alias{read.msatsi_summary_ext}
\alias{read_msatsi}
\alias{summary.msatsi_input}
\title{MSATSI I/O}
\usage{
read.msatsi_slboot_trpl(fi, ...)

read.msatsi_mat(fi, ...)

read.msatsi_slboot_tensor(fi, ...)

read.msatsi_summary(fi, ...)

read.msatsi_summary_ext(fi, ...)

read.msatsi_sat(fi, ...)

\method{summary}{msatsi_input}(object, ...)

\method{plot}{msatsi_input.summary}(x, ...)
}
\arguments{
\item{fi}{character; the filename}

\item{...}{additional parameters, usually passed to
\code{\link{read.table}},
\code{\link{print}},
\code{\link[R.matlab]{readMat}}, or
\code{\link{plot}}}

\item{x,object}{an object}
}
\description{
The functions here can be used
to read MSATSI's standardized
set of input and output files
into R.
}
\details{
Each function produces an object with a specific class, with at least
a \code{\link{print}} method

\code{\link{read.msatsi_mat}} uses \code{\link[R.matlab]{readMat}} to
read in from matlab's binary data format; it drops the header information,
which only includes creation date and system platform information.
}

