% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/msatsi.R
\name{read_msatsi}
\alias{print.slboot_tensor}
\alias{print.slboot_trpl}
\alias{read.slboot_tensor}
\alias{read.slboot_trpl}
\alias{read_msatsi}
\title{MSATSI output readers}
\usage{
read.slboot_trpl(fi, ...)

\method{print}{slboot_trpl}(x, ...)

read.slboot_tensor(fi, ...)

\method{print}{slboot_tensor}(x, ...)
}
\description{
MSATSI output readers
}

