% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msatsi.R
\name{read_msatsi}
\alias{read_msatsi}
\alias{read.msatsi_slboot_trpl}
\alias{read.msatsi_mat}
\alias{read.msatsi_slboot_tensor}
\alias{read.msatsi_summary}
\alias{read.msatsi_summary_ext}
\alias{read.msatsi_sat}
\title{MSATSI I/O}
\usage{
read.msatsi_slboot_trpl(fi, ...)

read.msatsi_mat(fi, ...)

read.msatsi_slboot_tensor(fi, ...)

read.msatsi_summary(fi, ...)

read.msatsi_summary_ext(fi, ...)

read.msatsi_sat(fi, ...)
}
\arguments{
\item{fi}{character; the filename}

\item{...}{additional parameters, usually passed to
\code{\link{read.table}}, or
\code{\link[R.matlab]{readMat}}}
}
\description{
The functions here can be used
to read MSATSI's standardized
set of input and output files
into R.
}
\details{
Each function produces an object with a specific class, with at least
a \code{\link{print}} method

\code{\link{read.msatsi_mat}} uses \code{\link[R.matlab]{readMat}} to
read in from matlab's binary data format; it drops the header information,
which only includes creation date and information on the original
system platform.
}
\seealso{
\code{\link{msatsi-methods}}
}
