% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/msatsi.R
\name{relative_stress}
\alias{.deviatoric_tensor}
\alias{relative_stress}
\alias{relative_stress.msatsi_slboot_tensor}
\title{Relative principal stress magnitudes and directions}
\usage{
relative_stress(x, ...)

.deviatoric_tensor(s, ...)

\method{relative_stress}{msatsi_slboot_tensor}(x, ...)
}
\arguments{
\item{x}{object to extract relative stress magnitudes from}

\item{...}{additional parameters to \code{\link{.deviatoric_tensor}}}

\item{s}{numeric; a vector of stress tensor components in the following order:
s11 s12 s13 s22 s23 s33}
}
\description{
This function finds the principal
stresses from the eigenvalues and eigenvectors of
the deviatoric stress tensor, \code{s}, defined as the stress tensor
minus the mean (or volumetric) stress.
}
\details{
Uses \code{\link{eigen}} to return the orientation
and magnitudes (in MPa -- check this) of the relative principal stresses.
}
\seealso{
\code{\link{read_msatsi}}
}

