% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spatial.R
\name{to_raster}
\alias{to_raster}
\alias{to_raster.msatsi_summary_ext}
\title{Convert objects to raster objects}
\usage{
to_raster(x, ...)

\method{to_raster}{msatsi_summary_ext}(x, ...)
}
\description{
Convert objects to raster objects
}

