% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{TP}
\alias{TP}
\title{Stereonet representation of trend and plunge}
\usage{
TP(ang.trend, ang.plunge, r = NULL, add = FALSE, pt.pch = 16,
  pt.col = "black", pt.cex = 0.8, relpos = c(0, 0), dressing = TRUE,
  markers = TRUE, ...)
}
\arguments{
\item{ang.trend}{numeric; trend or azimuth}

\item{ang.plunge}{numeric; plunch or dip}

\item{r}{numeric; the radius of the stereonet}

\item{add}{logical; should the points be added to the current device?}

\item{pt.pch, pt.col, pt.cex}{numeric; the symbol type(s) of the points,
the color(s) of the points, and the \code{cex}(s) of the points}

\item{relpos}{numeric; a vector giving the relative shift to apply to the TP projection}

\item{dressing}{logical; include projection markings?}

\item{markers}{logical; should markings for the projection be included? (irrelevant if \code{!dressing})}

\item{...}{additional arguments to \code{\link{points}}}
}
\description{
Uses the equal-area (lower) hemispherical, or \emph{Schmidt}, projection.
 From [1]:  The main advantage of this projection
 method is that it does not suffer from the areal distortion of the equal angle
 projection. This means, for example, that an area such as that enclosed by a circle
 of constant radius on the projection, represents the same amount of area on the
 reference sphere regardless of its position. The circle will not however represent a
 geometrically similar circle on the sphere unless it is positioned at the centre of the
 projection. In other words, areal relationships are preserved in this method while
 geometrical relationships are distorted; the opposite is true of the equal angle
 projection.
}
\details{
Uses \code{\link{complex}}, which recycles input vectors as needed
}
\examples{

par(mar=rep(0,4), oma=rep(0.2,4))
TP(runif(20, min = 0, max = 90),c(45,55), xlim=c(-2,2), ylim=c(-2,2))
TP(runif(20, min = 180, max = 270),c(45,55), pt.pch=1, add=TRUE)
TP(runif(20, min = 180, max = 270),c(45,55), pt.col='red', add=TRUE, relpos=c(2,0.5), markers=FALSE)

}
\references{
[1] Diederichs, Mark S (1990),
Dips: An Interactive and Graphical Approach to the Analysis of Orientation Based Data,
\url{https://www.rocscience.com/documents/pdfs/uploads/7672.pdf}
}

